e2e: merge black_box.mjs into support.mjs; return spawnSync result raw
Single support module instead of two; black_box.mjs only had four
helpers, all consumed exclusively by support.mjs. Drop the matching
js_library and the support → black_box dep edge.

runNodeScript no longer rebuilds a 4-field object — just returns
spawnSync's result. assertNodeOutput plucks the fields it cares about
before deepEqual.

BAZEL_TEST_INVOCATIONS=buildbuddy:91dda92f-e915-47bf-a133-49736b4166da
